continued on task logic
added more task logic and some PDO formatting, also added some javascript to make the the forms more dynamic. to do next: fix "completed" feature

+++ app/posts/delete.php
@@ -9,8 +9,10 @@ require __DIR__ . '/../autoload.php';
 //i want to add so that it compares the user ID with the logged in users ID before it deletes anything... to be continued.
 if (isset($_POST['taskIdToDelete'])) {
     $taskId = $_POST['taskIdToDelete'];
-    $statement = $database->prepare('DELETE FROM tasks WHERE id = :id');
+    $userId = $_SESSION['user']['id'];
+    $statement = $database->prepare('DELETE FROM tasks WHERE id = :id AND user_id = :user_id');
     $statement->bindParam(':id', $taskId, PDO::PARAM_INT);
+    $statement->bindParam(':user_id', $userId, PDO::PARAM_INT);
     $statement->execute();
 }
